requirements.txt
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
boto3==1.36.6
botocore==1.36.6
cffi==1.17.1
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.0
django-cors-headers==4.6.0
django-debug-toolbar==5.0.1
django-dotenv==1.4.2
django-htmx==1.21.0
django-redis==5.4.0
django-storages==1.14.4
djangorestframework==3.15.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==1.0.1
numpy==2.2.2
pandas==2.2.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
s3transfer==0.11.2
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
zope.interface==7.2
